requirements.txt
flask
transformers
datasets
peft
accelerate
mysql-connector-python
torch
tailwindcss
gunicorn
